+++ rust/examples-frontend/videoroom/webpack.config.js
@@ -1,0 +1,27 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const path = require('path');
+
+module.exports = {
+	mode   : 'development',
+	entry  : './src/index.ts',
+	module : {
+		rules : [
+			{
+				test    : /\.ts$/,
+				use     : 'ts-loader',
+				exclude : /node_modules/
+			}
+		]
+	},
+	resolve : {
+		extensions : [ '.ts', '.js' ]
+	},
+	output : {
+		filename : 'dist/bundle.js',
+		path     : path.resolve(__dirname, 'dist')
+	},
+	devServer : {
+		liveReload : false,
+		port       : 3001
+	}
+};