+++ rust/src/lib.rs
@@ -53,6 +53,7 @@ mod macros;
 mod messages;
 #[doc(hidden)]
 pub mod ortc;
+pub mod prelude;
 pub mod router;
 pub mod rtp_parameters;
 pub mod scalability_modes;
@@ -66,6 +67,12 @@ pub mod audio_level_observer {
     //! An audio level observer monitors the volume of the selected audio producers.
 
     pub use crate::router::audio_level_observer::*;
+}
+
+pub mod active_speaker_observer {
+    //! An active speaker observer monitors the speaking activity of the selected audio producers.
+
+    pub use crate::router::active_speaker_observer::*;
 }
 
 pub mod consumer {