+++ rust/src/messages.rs
@@ -1,3 +1,4 @@
+use crate::active_speaker_observer::ActiveSpeakerObserverOptions;
 use crate::audio_level_observer::AudioLevelObserverOptions;
 use crate::consumer::{
     ConsumerDump, ConsumerId, ConsumerLayers, ConsumerScore, ConsumerStats, ConsumerTraceEventType,
@@ -411,6 +412,30 @@ request_response!(
     RouterCreateAudioLevelObserverRequest {
         internal: RtpObserverInternal,
         data: RouterCreateAudioLevelObserverData,
+    },
+);
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub(crate) struct RouterCreateActiveSpeakerObserverData {
+    pub(crate) interval: u16,
+}
+
+impl RouterCreateActiveSpeakerObserverData {
+    pub(crate) fn from_options(
+        active_speaker_observer_options: &ActiveSpeakerObserverOptions,
+    ) -> Self {
+        Self {
+            interval: active_speaker_observer_options.interval,
+        }
+    }
+}
+
+request_response!(
+    "router.createActiveSpeakerObserver",
+    RouterCreateActiveSpeakerObserverRequest {
+        internal: RtpObserverInternal,
+        data: RouterCreateActiveSpeakerObserverData,
     },
 );