+++ rust/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-# Unreleased
+# 0.8.4
+
+* Add Active Speaker Observer to prelude
+* Fix consumers preventing producers from being closed (regression introduced in 0.8.3)
+
+# 0.8.3
+
+* prelude module containing traits and structs that should be sufficient for most basic mediasoup-based apps
+* Dominant Speaker Event (PR #603 by @SteveMcFarlin).
+
+### 0.8.2
 
 * Support for optional fixed port on transports